split Search and HopePage, and put a Welcome message on the Hompage, that also meantions the User's firt_name if logged in
